refactor: move all domain classes inside 'domain' package
