+++ app/src/main/java/com/codesoom/assignment/domain/dtos/TaskDTO.java
@@ -1,4 +1,4 @@
-package com.codesoom.assignment.dtos;
+package com.codesoom.assignment.domain.dtos;
 
 public class TaskDTO {
     private String task;
