+++ app/src/main/java/com/codesoom/assignment/domain/persistences/TaskRepository.java
@@ -1,6 +1,6 @@
-package com.codesoom.assignment.persistences;
+package com.codesoom.assignment.domain.persistences;
 
-import com.codesoom.assignment.entity.Task;
+import com.codesoom.assignment.domain.entity.Task;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
